+++ src/Core/Fields/FieldsFilter.php
@@ -15,6 +15,12 @@ namespace Bitrix24\SDK\Core\Fields;
 
 class FieldsFilter
 {
+    private const CRM_USER_FIELD_PREFIX = 'UF_CRM_';
+
+    private const SMART_PROCESS_FIELD_PREFIX = 'PARENT_ID_';
+
+    private const PRODUCT_USER_FIELD_PREFIX = 'PROPERTY_';
+
     /**
      * @param array<int, non-empty-string> $fieldCodes
      * @return array<int, non-empty-string>
@@ -23,7 +29,9 @@ class FieldsFilter
     {
         $res = [];
         foreach ($fieldCodes as $fieldCode) {
-            if (!str_starts_with($fieldCode, 'UF_CRM_') && !str_starts_with($fieldCode, 'PARENT_ID_')) {
+            if (!str_starts_with($fieldCode, self::CRM_USER_FIELD_PREFIX) &&
+                !str_starts_with($fieldCode, self::SMART_PROCESS_FIELD_PREFIX) &&
+                !str_starts_with($fieldCode, self::PRODUCT_USER_FIELD_PREFIX)) {
                 $res[] = $fieldCode;
             }
         }
@@ -39,7 +47,7 @@ class FieldsFilter
     {
         $res = [];
         foreach ($fieldCodes as $fieldCode) {
-            if (str_starts_with($fieldCode, 'UF_CRM_')) {
+            if (str_starts_with($fieldCode, self::CRM_USER_FIELD_PREFIX)) {
                 $res[] = $fieldCode;
             }
         }
@@ -55,7 +63,7 @@ class FieldsFilter
     {
         $res = [];
         foreach ($fieldCodes as $fieldCode) {
-            if (str_starts_with($fieldCode, 'PARENT_ID_')) {
+            if (str_starts_with($fieldCode, self::SMART_PROCESS_FIELD_PREFIX)) {
                 $res[] = $fieldCode;
             }
         }